moved dream specific blas calls to dream
