+++ SparseGrids/tsgHiddenExternals.hpp
@@ -44,7 +44,6 @@ extern "C" double dnrm2_(const int *N, const double *x, const int *incx);
 extern "C" void dgemv_(const char *transa, const int *M, const int *N, const double *alpha, const double *A, const int *lda, const double *x, const int *incx, const double *beta, const double *y, const int *incy);
 extern "C" void dtrsv_(const char *uplo, const char *trans, const char *diag, const int *N, const double *A, const int *lda, const double *x, const int *incx);
 extern "C" void dgemm_(const char* transa, const char* transb, const int *m, const int *n, const int *k, const double *alpha, const double *A, const int *lda, const double *B, const int *ldb, const double *beta, const double *C, const int *ldc);
-extern "C" void dtrsm_(const char *side, const char *uplo, const char* transa, const char* diag, const int *m, const int *n, const double *alpha, const double *A, const int *lda, const double *B, const int *ldb);
 #endif // Tasmanian_ENABLE_BLAS
 
 
@@ -75,26 +74,10 @@ namespace TasBLAS{
         char charT = 'T'; int blas_one = 1; double alpha = 1.0, beta = 0.0;
         dgemv_(&charT, &M, &N, &alpha, A, &M, x, &blas_one, &beta, y, &blas_one);
     }
-    inline void dtrsv_UTN(int N, const double A[], double B[]){ // solve A = U^{-T} B (A comes from cholUTU)
-        char charU = 'U', charT = 'T', charN = 'N'; int ione = 1;
-        dtrsv_(&charU, &charT, &charN, &N, A, &ione, B, &ione);
-    }
-    inline void dtrsv_UNN(int N, const double A[], double B[]){ // solve A = U^{-1} B (A comes from cholUTU)
-        char charU = 'U', charN = 'N'; int ione = 1;
-        dtrsv_(&charU, &charN, &charN, &N, A, &ione, B, &ione);
-    }
     // Level 3
     inline void dgemm(int M, int N, int K, double alpha, const double A[], const double B[], double beta, double C[]){
         char charN = 'N';
         dgemm_(&charN, &charN, &M, &N, &K, &alpha, A, &M, B, &K, &beta, C, &M);
-    }
-    inline void dtrsm_LUTN(int M, int N, const double A[], const double B[]){ // solve A = U^{-T} B (A comes from cholUTU)
-        char charL = 'L', charU = 'U', charT = 'T', charN = 'N'; double alpha = 1.0; int ione = 1;
-        dtrsm_(&charL, &charU, &charT, &charN, &M, &N, &alpha, A, &ione, B, &ione);
-    }
-    inline void dtrsm_LUNN(int M, int N, const double A[], const double B[]){ // solve A = U^{-1} B (A comes from cholUTU)
-        char charL = 'L', charU = 'U', charN = 'N'; double alpha = 1.0; int ione = 1;
-        dtrsm_(&charL, &charU, &charN, &charN, &M, &N, &alpha, A, &ione, B, &ione);
     }
 #else
     // non optimal BLAS subroutines, in case there is no BLAS available
@@ -134,54 +117,7 @@ namespace TasBLAS{
             for(int j=1; j<N; j++) y[i] += A[i*M + j] * x[j];
         }
     }
-    inline void dtrsv_UTN(int N, const double A[], double B[]){ // solve A = U^{-T} B (A comes from cholUTU)
-        B[0] /= A[0];
-        for(int i=1; i<N; i++){
-            for(int j=0; j<i; j++) B[i] -= A[i*N + j] * B[i];
-            B[i] /= A[i*N + i];
-        }
-    }
-    inline void dtrsv_UNN(int N, const double A[], double B[]){ // solve A = U^{-1} B (A comes from cholUTU)
-        for(int i=N-1; i>=0; i--){
-            B[i] /= A[(i+1)*N-1];
-            for(int j=i-1; j>=0; j--) B[j] -= A[i*N + j] * B[i];
-        }
-    }
-    inline void dtrsm_LUTN(int M, int N, const double A[], double B[]){ // solve A = U^{-T} B (A comes from cholUTU)
-        //char charL = 'L', charL = 'U', charL = 'T', charN = 'N'; double alpha = 1.0; int ione = 1;
-        //dtrsm_(&charL, &charU, &charT, &charN, &M, &N, &alpha, A, &ione, B, &ione);
-        for(int b=0; b<N; b++){ // for each column of B
-            B[b*M] /= A[0];
-            for(int i=1; i<M; i++){
-                for(int j=0; j<i; j++) B[b*M + i] -= A[i*M + j] * B[b*M + i];
-                B[b*M + i] /= A[i*M + i];
-            }
-        }
-    }
-    inline void dtrsm_LUNN(int M, int N, const double A[], double B[]){ // solve A = U^{-1} B (A comes from cholUTU)
-        //char charL = 'L', charL = 'U', charN = 'N'; double alpha = 1.0; int ione = 1;
-        //dtrsm_(&charL, &charU, &charN, &charN, &M, &N, &alpha, A, &ione, B, &ione);
-        for(int b=0; b<N; b++){ // for each column of B
-            for(int i=M-1; i>=0; i--){
-                B[b*M+i] /= A[(i+1)*M-1];
-                for(int j=i-1; j>=0; j--) B[b*M + j] -= A[i*M + j] * B[b*M + i];
-            }
-        }
-    }
 #endif // Tasmanian_ENABLE_BLAS
-
-    // used to compute anisotropic weights (at most 2 * dims X 2 * dims, not worth adding lapack dependence for such small matrix)
-    inline void cholUTU(int N, double A[], int lda = -1){
-        if (lda == -1) lda = N;
-        for(int i=0; i<N; i++){
-            for(int k=0; k<i; k++) A[i*lda + i] -= A[i*lda +k]*A[i*lda +k];
-            A[i*lda + i] = sqrt(A[i*lda + i]);
-            for(int j=i+1; j<N; j++){
-                for(int k=0; k<i; k++) A[j*lda + i] -= A[i*lda +k]*A[j*lda +k];
-                A[j*lda + i] /= A[i*lda + i];
-            }
-        }
-    }
 
     inline void setzero(int N, double *A){ std::fill(A, A + N, 0.0); } // fill A with N zeros
     inline void setzero(int N, int *A){ std::fill(A, A + N, 0); } // fill A with N zeros
